Added osquery-perf support.
